requirements.txt
pathway
python-dotenv
litellm
pandas
openai
gdown
